requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.7
deepdiff==8.0.1
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
emoji==2.12.1
fastapi==0.114.1
fastapi-cli==0.0.5
filetype==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.2
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langchain==0.2.16
langchain-community==0.2.16
langchain-core==0.2.39
langchain-ollama==0.1.3
langchain-openai==0.1.24
langchain-text-splitters==0.2.4
langchain-unstructured==0.1.2
langdetect==1.0.9
langsmith==0.1.118
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mdurl==0.1.2
multidict==6.1.0
mypy-extensions==1.0.0
narwhals==1.7.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.3.3
openai==1.45.0
orderly-set==5.2.2
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pdf2image==1.17.0
pillow==10.4.0
protobuf==5.28.1
psutil==6.0.0
pyarrow==17.0.0
pydantic==2.9.1
pydantic_core==2.23.3
pydeck==0.9.1
Pygments==2.18.0
pypdf==4.3.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
python-oxmsg==0.0.1
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==3.9.7
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.8.1
rpds-py==0.20.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.34
starlette==0.38.5
streamlit==1.38.0
tabulate==0.9.0
tenacity==8.5.0
tiktoken==0.7.0
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
unstructured==0.15.10
unstructured-client==0.24.1
urllib3==2.2.3
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.24.0
websockets==13.0.1
wrapt==1.16.0
yarl==1.11.1
